requirements.txt
numpy
pandas
seaborn
streamlit>=1.14
mysql-connector-python
google-api-python-client
SQLAlchemy
plotly-express